ユーザ退会　Postモデル、ユーザコントローラ修正

+++ app/Http/Controllers/UsersController.php
@@ -41,9 +41,13 @@ class UsersController extends Controller
 
     public function destroy($id)
     {
-        $user = \Auth::user();
-        $user->delete();
-        return back();
+
+        $user = User::findOrFail($id);
+        if (\Auth::id() === $user->id) {
+            $user->delete();
+        }
+            return back();
+        
     }
 
 }
